Fix boolean issue in Java bindings.
This is due to the fact that jni::sys::jboolean is just a typedef to u8.
When I was setting the the boolean value, I was using the JValue conversion.
It was converting it erroneously to a u8, not a boolean. The rest is just
the JVM having a very bad time with it.