Working my way on all the conversions.
There are still some hard to express conversions to do, it's
way too late and I'm starting to lose faith in the cpp
generator. But as Scarlett O'Hara said, "after all,
tomorrow is another day".

And to you Rhett, we frankly don't give a damn
about you. Scarlett was the protagonist we seek after,
not you.

Note: I just rewatched GwtW, so bear with me.